Final DB

+++ checked-items/checked-items.php
@@ -22,6 +22,7 @@
                 <button class = "navButton" onclick="document.location='\\checked-items\\checked-items.php'">Checkedout Items</button>
                 <button class = "navButton" onclick="document.location='\\fees\\fees.php'">Fees</button>
                 <button class = "navButton" onclick="document.location='\\account\\account.php'">Account</button>
+                <button class ="navButton" onclick ="document.location='\\index.php'">Logout</button>
         </div>
 </div>
 
@@ -43,12 +44,12 @@
     session_start();
     $uid = $_SESSION["ID"];
 
-    $server = 'localhost';
-    $username = 'root'; 
-    $password = 'root1234'; 
-    $database = 'library'; 
+    $servername = "spring2024-gp9-library-azure.mysql.database.azure.com";
+    $username = "gp9library";
+    $password = "Securewalls2";
+    $dbname = "library";
 
-    $conn = mysqli_connect($server, $username, $password, $database);
+    $conn = mysqli_connect($servername, $username, $password, $dbname);
 
 
     $checkedItems = [];
@@ -107,7 +108,7 @@
       }
 
       #delete user's checked entry
-      $query = "DELETE FROM checkedout_items WHERE item_id = '$id'";
+      $query = "DELETE FROM checkedout_items WHERE item_id = '$id' and account_id = '$uid'";
       mysqli_query($conn, $query);
 
       #check if there is hold on the item
